Updated pages to match style
Updated all recipe pages to match the style and layout of the index and spaghetti sauce pages. Removed the h1 element and moved the style rules to the headerTitle element. Still need to fix the navbar links for garlic bread and spaghetti page. And add a link to the title itself.