Title linked to index page
Linked the tile of each page to the Index and moved around the anchor related style rules to clean up the css style sheet. Noticed bug with some of the page elements having a space on the right side/ some elements overflowing off the page, still needs to be fixed.